move code specific to thumbnailing to thumbnailer
